refactor: 직선 레일 시리얼 등록 로직을 `StraightSerialWriteService`로 이동합니다.
