+++ src/main/java/baekgwa/suhoserver/domain/project/service/ProjectWriteService.java
@@ -8,7 +8,6 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
-import java.util.stream.LongStream;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -21,7 +20,6 @@ import baekgwa.suhoserver.model.project.branch.entity.ProjectBranchEntity;
 import baekgwa.suhoserver.model.project.branch.repository.ProjectBranchRepository;
 import baekgwa.suhoserver.model.project.project.entity.ProjectEntity;
 import baekgwa.suhoserver.model.project.project.repository.ProjectRepository;
-import baekgwa.suhoserver.model.project.straight.serial.entity.ProjectStraightSerialEntity;
 import baekgwa.suhoserver.model.project.straight.serial.repository.ProjectStraightSerialRepository;
 import baekgwa.suhoserver.model.project.straight.straight.entity.ProjectStraightEntity;
 import baekgwa.suhoserver.model.project.straight.straight.repository.ProjectStraightRepository;
@@ -161,9 +159,6 @@ public class ProjectWriteService {
 	}
 
 	/**
-
-	 */
-	/**
 	 * 프로젝트에 할당된 특정 직선레일 할당 해제
 	 * @param projectStraightId 프로젝트에 할당된 직선레일 PK
 	 * @return 추가로 삭제할 Straight Info PK
@@ -224,17 +219,6 @@ public class ProjectWriteService {
 
 		// 2. 업데이트 처리
 		findProjectBranch.patchProjectBranch(dto.getTotalQuantity());
-	}
-
-	@Transactional
-	public void registerProjectStraightSerial(List<ProjectStraightEntity> projectStraightList) {
-		List<ProjectStraightSerialEntity> projectStraightSerialList = projectStraightList.stream()
-			.flatMap(ps -> LongStream.rangeClosed(1L, ps.getTotalQuantity())
-				.mapToObj(seq -> ProjectStraightSerialEntity.of(ps, seq))
-			)
-			.toList();
-
-		projectStraightSerialRepository.saveAll(projectStraightSerialList);
 	}
 
 	private void validateDuplicationStraight(
